Add Loading as isoleted Compose
