+++ SDKAndroid/src/main/java/app/cybrid/sdkandroid/ui/Theme/Font.kt
@@ -10,3 +10,7 @@ val robotoFont = FontFamily(
     Font(R.font.roboto_medium, weight = FontWeight.Medium),
     Font(R.font.roboto_bold, weight = FontWeight.Bold)
 )
+
+val interFont = FontFamily(
+    Font(R.font.inter_regular, weight = FontWeight.Normal)
+)
