+++ SDKAndroid/src/main/res/values-es/strings.xml
@@ -35,6 +35,8 @@
     <string name="accounts_view_loading_text">Cargando</string>
     <string name="accounts_view_list_header_asset">Activo</string>
     <string name="accounts_view_balance_title">Valor de la cuenta</string>
+    <string name="accounts_view_balance_available_title">Balance para transacciones</string>
+    <string name="accounts_view_pending_deposit_label"> pending deposit</string>
     <string name="accounts_view_list_header_asset_sub">Precio en el mercado</string>
     <string name="accounts_view_list_header_balance">Balance</string>
     <string name="accounts_view_trades_holding_title">Respaldados</string>
@@ -71,5 +73,8 @@
     <string name="bank_accounts_view_done_button">Continuar</string>
     <string name="bank_accounts_view_error_text">Ha ocurrido un error, vuelve a intentarlo.</string>
     <string name="bank_accounts_view_error_button">Listo</string>
+
+    <!-- Transfer Component -->
+    <string name="transfer_view_component_loading_view_title">Cargando</string>
 
 </resources>
