+++ SDKAndroid/src/main/res/values/colors.xml
@@ -20,12 +20,15 @@
     <color name="modal_sub_title_refresh_color">#FF5252</color>
     <color name="modal_divider">#79747E</color>
 
+    <!-- Accounts View -->
     <color name="accounts_view_balance_title">#616161</color>
     <color name="accounts_view_trades_holdings_title">#424242</color>
     <color name="accounts_view_trades_separator">#E0E0E0</color>
     <color name="accounts_view_trades_buy">#4CAF50</color>
     <color name="accounts_view_trades_sell">#757575</color>
     <color name="account_trade_detail_content_item_color">#424242</color>
+    <color name="accounts_balance_available_trade_color">#636366</color>
+    <color name="accounts_pending_deposit_color">#5AC8FA</color>
 
     <color name="accent_blue">#007AFF</color>
