+++ SDKAndroid/src/main/res/values/strings.xml
@@ -35,6 +35,8 @@
     <!-- Accounts View Strings -->
     <string name="accounts_view_loading_text">Loading</string>
     <string name="accounts_view_balance_title">Account Value</string>
+    <string name="accounts_view_balance_available_title">Available to Trade</string>
+    <string name="accounts_view_pending_deposit_label"> pending deposit</string>
     <string name="accounts_view_list_header_asset">Asset</string>
     <string name="accounts_view_list_header_asset_sub">Market Price</string>
     <string name="accounts_view_list_header_balance">Balance</string>
@@ -72,5 +74,8 @@
     <string name="bank_accounts_view_done_button">Continue</string>
     <string name="bank_accounts_view_error_text">An error has occurred, try again.</string>
     <string name="bank_accounts_view_error_button">Done</string>
+
+    <!-- Transfer Component -->
+    <string name="transfer_view_component_loading_view_title">Loading</string>
 
 </resources>
